Updated PatternList tests to use the newly renamed EditorForm component. Refactored tests to remove obsolete props
